chore(Board): move Board and BoardContainer
